TASK: Move SubtreeInterface
